Add wrapper for operation operation_scenarios
Extends existing ZEN-garden functionalities to enable operation-only
simulations after capacity-planning simulations. Pilots a new "wrapper"
format, in which changes are programmed as a wrapper that runs
ZEN-garden in specific configurations. The wrapper can
be accessed via the command line, making it a practical extensions of
the model. Wrappers avoid clutter in the core model code.

The new wrapper allows users to specify a model for which capacities
were optimized. Then, the added capacities are transferred to existing
capacities in a copy of the model. This new model is then run in
operation-only format. Users may specify new scenario files for the
operation-only simulations.