Remove operation-only-phase features
The operation only phase (commits ab310ce, 39e3dcd, 8534302)
functionality has now been replaced by the new operation wrapper.
The wrapper is a cleaner implementation of the same functionality since
it enables increased flexibility (e.g. to change the number of
hours in the simulation) and declutters the core zen-garden code.
This commit removes the old functionality to clean up the code.